+++ fairscale/nn/pipe/async_schedule.py
@@ -11,15 +11,64 @@ from typing import Dict, Iterable, List, Optional, Tuple
 
 import torch
 from torch import Tensor, nn
+from torch.autograd.profiler import record_function
 from torch.distributed import ProcessGroup
 
 from fairscale.nn.model_parallel import get_pipeline_parallel_ranks
 
+from .checkpoint import Checkpointing
 from .messages import Transport
 from .microbatch import Batch
-from .multiprocess_pipeline import create_task
-from .skip.tracker import SkipTrackerThroughPotals
-from .types import EVENT_LOOP_QUEUE, PipeMessage, Tensors
+from .skip.tracker import SkipTrackerThroughPotals, use_skip_tracker
+from .types import EVENT_LOOP_QUEUE, PipeMessage, TensorOrTensors, Tensors
+from .worker import Task
+
+
+def create_task(
+    checkpoint_stop: int,
+    i: int,
+    j: int,
+    batch: Batch,
+    partition: nn.Sequential,
+    skip_trackers: List[SkipTrackerThroughPotals],
+) -> Task:
+    # Determine whether checkpointing or not.
+    if i < checkpoint_stop:
+
+        def function(
+            input: TensorOrTensors,
+            partition: nn.Sequential = partition,
+            skip_tracker: SkipTrackerThroughPotals = skip_trackers[i],
+            chunk_id: int = i,
+            part_id: int = j,
+        ) -> TensorOrTensors:
+            with use_skip_tracker(skip_tracker), record_function("chunk%d-part%d" % (chunk_id, part_id)):
+                ret = partition(input)
+                # We do a check here because the backtrace from the checkpoint backward code path
+                # is very hard to make sense. It would be much easier to check earlier at this point.
+                assert type(ret) is not list, "Only Tensor or Tuple of Tensor output is supported"
+                return ret
+
+        chk = Checkpointing(function, batch)
+        task = Task(None, compute=chk.checkpoint, finalize=chk.recompute)
+        del function, chk  # TODO(tom) maybe remove
+
+    else:
+
+        def compute(
+            batch: Batch = batch,
+            partition: nn.Sequential = partition,
+            skip_tracker: SkipTrackerThroughPotals = skip_trackers[i],
+            chunk_id: int = i,
+            part_id: int = j,
+        ) -> Batch:
+            with use_skip_tracker(skip_tracker), record_function("chunk%d-part%d" % (chunk_id, part_id)):
+                return batch.call(partition)
+
+        task = Task(None, compute=compute, finalize=None)
+        del compute  # TODO(tom) maybe remove
+
+    return task
 
 
 @dataclass(frozen=True)
